requirements.txt
pytesseract
Pillow
pandas
watchdog
opencv-python
pdf2image
pdfplumber